Remove shadowed _baseDemographics and _demographicsPercentile properties
_baseDemographics wasn't doing anything that typescript isn't already doing by verifying that the entity is a typed record. Refactored to use the direct baseDemographics property. Done in `createTown` and `createTownBiome`

Similarly `updateDemographics` was setting the baseDemographics property unnecessarily.

Finally, we were calculating demographics percentages in a few different places. This adds a getter that always calculates the percentage, so may result in a few more additional calculations at the price of doing it in just a single place. May need to add memoization if performance takes a hit, but this seems unlikely.