Refactor race (#704)
* Adding macos droppings to .gitignore

* Refactor fetchProfessionChance and adding unit tests

- Extracting filtering work into separate methods
- Moving some math-y stuff to separate file and adding unit tests

* Updating typescript linting packages

Running linters caused warning in console that this version of the typescript eslint parser was not compatible with the version of typescript used. This resolves that warning

* Removing unused and unnecessary file

* Refactor fetchRace and adding tests

- Instead of a 'private' _baseDemographics shadowed varible, this fetches race data directly and uses it with a shared math function

* Remove shadowed _baseDemographics and _demographicsPercentile properties

_baseDemographics wasn't doing anything that typescript isn't already doing by verifying that the entity is a typed record. Refactored to use the direct baseDemographics property. Done in `createTown` and `createTownBiome`

Similarly `updateDemographics` was setting the baseDemographics property unnecessarily.

Finally, we were calculating demographics percentages in a few different places. This adds a getter that always calculates the percentage, so may result in a few more additional calculations at the price of doing it in just a single place. May need to add memoization if performance takes a hit, but this seems unlikely.

* Moving race generation higher in order

* re-adding percentile setter

* Getting rid of final _demographicPercentiles

* _demographicPercentile -> demographicPercentile

* updating signature of refactored fn