Added a Dockerfile to run EEEG in a container
This is an initial Dockerfile which will allow building the app as a Docker container. It's pretty basic; just installs the requisite dependencies and copies a list of source files into the image.
    
To build the Docker image, install Docker and run:
    docker build --tag=eeeg:latest .

To run the image, run:
    docker run -p 8080:80 eeeg:latest
and visit either http://localhost:8080 or the IP address which Docker assigns dynamically to the container (which probably looks like 172.17.0.1).